requirements.txt
asgiref==3.2.10
boto3==1.16.9
botocore==1.19.9
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-environ==0.4.5
djangorestframework==3.12.1
idna==2.10
jmespath==0.10.0
mysql-connector-python==8.0.22
pdfy==1.0.50
protobuf==3.13.0
PyChromeDevTools==0.1
pychromepdf==1.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
selenium==3.141.0
dnspython
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
websocket-client==0.57.0
